+++ internal/usecases/get_residence/deps.go
@@ -11,4 +11,5 @@ import (
 //go:generate mockgen -source $GOFILE -destination deps_test.go -package ${GOPACKAGE}
 type ResidenceRepo interface {
 	GetResidence(ctx context.Context, id models.ID) (*models.Residence, error)
+	GetPerson(ctx context.Context, id models.ID) (*models.Person, error)
 }